fix: rename buildComponentHierarchy file
