+++ packages/vue/src/FlareErrorBoundary.ts
@@ -2,7 +2,7 @@ import { flare } from '@flareapp/js';
 import type { ComponentPublicInstance } from 'vue';
 import { defineComponent, onErrorCaptured, ref } from 'vue';
 
-import { buildComponentHierarchy } from './componentHierarchy';
+import { buildComponentHierarchy } from './buildComponentHierarchy';
 import { convertToError } from './convertToError';
 import { getComponentName } from './getComponentName';
 import { FlareVueContext } from './types';
