+++ packages/vue/src/buildComponentHierarchy.ts
